Server_Django/requirements.txt
Django
dj_database_url
dotenv
mysqlclient
PyJWT
django-cors-headers